update MD convert bin script
