+++ bin/convert-MD.js
@@ -4,8 +4,8 @@ import convert from '../lib/convert/MD.js';
 import program from 'commander';
 
 program
-.arguments(`<inputPath> <outputPath> [mappingsPath]`)
-.usage(`convert-md <inputPath> <outputPath> [mappingsPath]`)
+.arguments(`<inputPath> <outputPath>`)
+.usage(`convert-md <inputPath> <outputPath>`)
 .action(convert);
 
 program.parse(process.argv);
