+++ lib/convert/MD.js
@@ -1,5 +1,5 @@
 import createSpinner from 'ora';
-import fs from 'fs-extra';
+import fs            from 'fs-extra';
 import readTSV       from '../utilities/readTSV.js';
 import writeNDJSON   from '../utilities/writeNDJSON.js';
