+++ lib/import/MD.js
@@ -1,167 +1,108 @@
-/**
- * Imports the MD entries into the ALTLab database.
- */
+import createSpinner from 'ora';
+import readNDJSON    from '../utilities/readNDJSON.js';
+import writeNDJSON   from '../utilities/writeNDJSON.js';
 
-import createHomographKey from '../utilities/createHomographKey.js';
-import createMDKey        from '../utilities/createMDKey.js';
-import Index              from '../utilities/DatabaseIndex.js';
-import loadEntries        from '../utilities/loadEntries.js';
-import saveDatabase       from '../utilities/saveDatabase.js';
+function createKey(sro, pos) {
+  sro = sro.replace(/\s+/gu, `_`);
+  return `${ sro }@${ pos }`;
+}
 
 /**
- * A class for an ALTLab database entry.
+ * A class representing a database index.
+ * @extends Map
  */
-class DBEntry {
+class DatabaseIndex extends Map {
 
   /**
-   * Create a new ALTLab database entry from an MD entry.
-   * @param {Object} mdEntry The MD entry to create the database entry from.
+   * Create a new database index. Assumes all entries unique SRO headword + POS combinations.
+   * @param {Array} entries An array of database entries to index.
    */
-  constructor(mdEntry) {
+  constructor(entries) {
 
-    mdEntry.matched = true;
+    super();
 
-    this.dataSources = {
-      MD: mdEntry,
-    };
+    for (const entry of entries) {
+      this.add(entry);
+    }
 
-    this.lemma = {
-      sro: mdEntry.mapping?.lemma_CW ?? mdEntry.lemma.md,
-    };
+  }
 
-    this.senses = mdEntry.senses
-    .map(sense => Object.assign({}, sense, { dataSource: `MD` }));
+  /**
+   * Adds an entry to the index.
+   * @param {Object} entry The entry to add to the index.
+   */
+  add(entry) {
 
+    const key           = createKey(entry.lemma.sro, entry.pos);
+    const existingEntry = this.get(key);
+
+    if (existingEntry) {
+
+      if (Array.isArray(existingEntry)) {
+        existingEntry.push(entry);
+        return this;
+      }
+
+      return this.set(key, [existingEntry, entry]);
+
+    }
+
+    return this.set(key, entry);
+
+  }
+
+  /**
+   * Removes an entry from the index.
+   * @param  {String} key The key of the entry to remove from the index.
+   * @return {Map}
+   */
+  remove(key) {
+    return this.delete(key);
   }
 
 }
 
 /**
  * Imports the MD entries into the ALTLab database.
- * @param  {String}  mdPath       The path to the MD NDJSON file.
- * @param  {String}  databasePath The path to the ALTLab database.
- * @return {Promise}              Returns a Promise that resolves to the Array of ALTLab database entries.
+ * @param  {String} mdPath
+ * @param  {String} dbPath
  */
-/* eslint-disable max-statements */
-export default async function importMD(mdPath, databasePath, { silent = false } = {}) {
+export default async function importMD(mdPath, dbPath) {
 
-  const db             = await loadEntries(databasePath);
-  const md             = await loadEntries(mdPath);
-  let   entriesAdded   = 0;
-  let   entriesDeleted = 0;
+  const readDatabaseSpinner = createSpinner(`Reading databases.`).start();
 
-  const primaryIndex = new Index(db, (entry, index) => createHomographKey(entry.lemma.sro, index));
+  const mdEntries      = await readNDJSON(mdPath);
+  const dbEntries      = await readNDJSON(dbPath);
+  const originalDBSize = dbEntries.length;
+  const   entriesUpdated = 0;
+  const   entriesAdded   = 0;
 
-  const cwIndex = new Index(db, ({ dataSources: { CW: cwEntry } }) => {
-    if (cwEntry) return `${ cwEntry.lemma.plains }:${ cwEntry.definition }`;
+  readDatabaseSpinner.succeed(`Databases read into memory.`);
+
+  const indexSpinner = createSpinner(`Indexing database.`).start();
+  const dbIndex      = new DatabaseIndex(dbEntries);
+
+  indexSpinner.succeed(`Database indexed.`);
+
+  for (const mdEntry of mdEntries) {
+
+
+  }
+
+  const writeSpinner = createSpinner(`Writing entries to database file.`).start();
+  const entries      = Array.from(dbIndex.values()).flat();
+
+  await writeNDJSON(dbPath, entries);
+
+  writeSpinner.succeed(`Entries written to database file.`);
+
+  console.info(`\n`);
+  console.table({
+    'Size of database prior to import:': originalDBSize,
+    'Entries added:':                    entriesAdded,
+    'Entries updated:':                  entriesUpdated,
+    'Net change in database size:':      entries.length - originalDBSize,
+    'Size of database after import:':    entries.length,
   });
-  const cwKeys = Array.from(cwIndex.keys());
-
-  const mdIndex = new Index(db, ({ dataSources: { MD: mdEntry } }) => {
-    if (mdEntry) return createMDKey(mdEntry);
-  });
-
-  // Keep track of which existing MD subentries in the database have been
-  // matched to an entry in the current MD database being imported.
-  for (const [, mdEntry] of mdIndex) {
-    mdEntry.matched = false;
-  }
-
-  let unmatchedMappings = 0;
-
-  // Match each entry in the MD database to an existing ALTLab entry,
-  // or create a new entry in the ALTLab database.
-  for (const mdEntry of md) {
-
-    // attempt to get ALTLab entry by matching its MD subentry
-    const mdKey   = createMDKey(mdEntry);
-    let   dbEntry = mdIndex.get(mdKey);
-
-    // attempt to get ALTLab entry by using the MD > CW mapping
-    if (!dbEntry && mdEntry.mapping) {
-
-      // attempt to find an entry with the full lemma + definition
-      const { definition_CW, lemma_CW } = mdEntry.mapping;
-      dbEntry = cwIndex.get(`${ lemma_CW }:${ definition_CW }`);
-
-      // if there's still no match, attempt to get ALTLab entry using just the lemma
-      if (!dbEntry) {
-        const matchedKey = cwKeys.find(key => key.startsWith(lemma_CW));
-        dbEntry          = cwIndex.get(matchedKey);
-      }
-
-      // no match found
-      if (!dbEntry) unmatchedMappings++;
-
-    }
-
-    // If there's a matching ALTLab entry, update it.
-    if (dbEntry) {
-      mdEntry.matched        = true;
-      dbEntry.dataSources.MD = mdEntry;
-      continue;
-    }
-
-    // If there's not a matching ALTLab entry, create one
-    // and add it to indexes.
-    const newEntry = new DBEntry(mdEntry);
-    primaryIndex.add(newEntry);
-    cwIndex.add(newEntry);
-    mdIndex.add(newEntry);
-    entriesAdded++;
-
-  }
-
-  // clean up MD subentries
-  for (const [key, dbEntry] of mdIndex) {
-
-    const mdEntry = dbEntry.dataSources.MD;
-
-    if (mdEntry.matched) {
-      // remove unnecessary properties from subentry
-      delete mdEntry.key;
-      delete mdEntry.matched;
-    } else {
-      // remove MD data sources that were never matched
-      // (this means they no longer exist in the original MD database / mappings table)
-      delete dbEntry.dataSources.MD;
-      mdIndex.remove(key);
-    }
-
-  }
-
-  // Set the "key" property of any entries that are missing it.
-  // AND Remove database entries without data sources.
-  for (const [key, dbEntry] of primaryIndex) {
-
-    if (Object.keys(dbEntry.dataSources).length) {
-      dbEntry.key = key;
-      continue;
-    }
-
-    primaryIndex.remove(key);
-    entriesDeleted++;
-
-  }
-
-  const entries = Array.from(primaryIndex.values());
-
-  if (!silent) {
-    /* eslint-disable sort-keys */
-    console.info(`\n`);
-    console.table({
-      'Size of database prior to import': db.length,
-      'Entries to import':                md.length,
-      'Entries deleted':                  entriesDeleted,
-      'Entries added':                    entriesAdded,
-      'Mappings without a match':         unmatchedMappings,
-      'Net change in database size':      entries.length - db.length,
-      'Size of database after import':    entries.length,
-    });
-  }
-
-  await saveDatabase(databasePath, entries, { silent });
-
-  return entries;
 
 }
